Amount 금액 관련 Integer -> Long 변경

+++ Moabayo_Client_Bank/src/main/java/com/sboot/moabayo/controller/BankController.java
@@ -236,7 +236,7 @@ public class BankController {
                                HttpServletRequest req) {
     	
     	
-    	Integer amount = (Integer) session.getAttribute("paid_amount");
+    	Long amount = (Long) session.getAttribute("paid_amount");
         KakaoApproveResponse approve = kakaoPayService.approve(amount, pgToken, tidFromQuery);
 
         // 승인 성공 → 완료 페이지로
