+++ Moabayo_Client_Bank/src/main/java/com/sboot/moabayo/vo/AccountTxMeta.java
@@ -6,7 +6,7 @@ import lombok.Data;
 //거래 로그에 같이 저장할 선택적 정보들
 @Data @Builder
 public class AccountTxMeta {
- private Integer approvedAmount; // 없으면 메서드 인자 amount 사용
+ private Long approvedAmount; // 없으면 메서드 인자 amount 사용
  private String approvedNum;
  private String accountType; // 없으면 애노테이션 type 사용
  private String category;
